fix(money-integrity): close Decimal/rounding gaps found in review
- model_dump(mode="json") in chat's orchestrator and import_excel so
  Decimal line amounts serialize instead of crashing tool calls/imports
- Bound Money/Rate to their column precision on both ends and split out
  CalculatedMoney for read-time aggregates (estimated_local_cap,
  dashboard/conversion sums) that aren't stored in a fixed-precision column
- Add migration 000020 to reconcile allocations that independent rounding
  in 000019 pushed past their report-line or conversion cap
- Also includes unrelated staged changes (OpenSpec proposals for AI
  provider SSRF, spreadsheet hardening, chat prompt-injection provenance,
  users tenant-authz/internal-lookup hardening; e2e auth/money-precision
  specs; doc updates) per explicit instruction to commit as staged

+++ services/budget/app/crud/report_line_crud.py
@@ -1,4 +1,5 @@
 from datetime import date
+from decimal import Decimal
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from app.models.report import ReportLineModel
@@ -11,7 +12,7 @@ async def create_report_line(
     report_id: UUID,
     budget_line_id: UUID,
     description: str,
-    amount: float,
+    amount: Decimal,
     expense_date: date,
     extra_fields: dict | None = None,
 ) -> ReportLineModel:
@@ -51,7 +52,7 @@ async def update_report_line(
     session: AsyncSession,
     report_line: ReportLineModel,
     description: str | None = None,
-    amount: float | None = None,
+    amount: Decimal | None = None,
     expense_date: date | None = None,
     extra_fields: dict | None = None,
 ) -> ReportLineModel:
